Refactoring templates
