+++ src/structuraltools/aci/resources/sectional_strength_templates_processed.json
@@ -1,0 +1,1 @@
+{}
